move: api-libにkeitaisoの内容をコピー
